PycharmProjects/cabinetsense-chatbot/requirements.txt
langchain
openai
chromadb
fastapi
uvicorn
python-dotenv
pydantic
